define some things for compatibility with the nypr-ui/hero component
`nypr-ui/hero` isn't quite as compatible as I thoiught. it relies on a 
few things defined in `nypr-ui` which are hard to split out or import in 
an isolated way.

this mixin defines those things (a mixin and two variables) in a way 
that shouldn't interfere with the rest of this app.